Use Spatial Data to Filter Point Data

Use a geospatial polygon to identify or remove points located outside the polygon. This function also includes an option buffer argument which can be used to add a specific distance buffer around the polygon prior to filtering.

nw.filterspatial(
  points,
  polygon,
  mode,
  buffer = NULL,
  buffer_units = NULL,
  buffer_output = NULL,
  proj = NULL,
  output = NULL
)

Arguments

points

sf; A spatial feature containing point data.

polygon

sf; A spatial polygon representing the area for which you want to include datafrom.

mode

"flag" or "remove"; A character string defining if the user wants the identified nesting attempts to be flagged with "FLAGGED" in a new column. Or removed from the dataset.

buffer

numeric; Optional distance to buffer around the polygon

buffer_units

"km" or "mi"; A character string defining the buffer distance units as either kilometers or miles.

buffer_output

logical; Optionally export the generated buffered polygon to the global enviornment.

proj

a PROJ.4 sting; Optional definition for map projection. Defaults to Lambert Conformal Conic.

output

character; An optional character string to custom name the output spatial dataframe.

Value

sf; A spatial dataframe

Examples