Preform Common Cleaning Procedures to NestWatch Data

NestWatch data are collected by volunteer participants (researchers & the public) and are known to contain some errors. This function provides several common procedures that NestWatch staff have identified to help the user clean NestWatch data for analysis. The user should consider their research objectives/species life histories and read the full documentation below and in the vignette.

nw.cleandata(data, mode, methods = NULL, output = NULL)

Arguments

data

dataframe; A dataframe containing NestWatch data.

mode

"flag" or "remove"; A character string defining if the user wants the identified nesting attempts to be flagged with "FLAGGED" in a new column. Or removed from the dataset.

methods

character vector; A vector containing any of the letters a:j, or "all", identifying each cleaning procedure to be conducted on the data. See Details below and the vignette for details on each procedure.

output

character; An optional character string to custom name the output dataframe.

Value

a dataframe

Details

NestWatch data in the database should be considered "raw" and the user should consider their system and research objectives when deciding what cleaning procedures to conduct. This function contains 10 cleaning procedures a:j.

Cleaning Procedure Details:

a: Flag/remove attempts with Species.Name is Brown-headed Cowbird. Brown-headed Cowbirds are nest parasites and do not create their own nests. This is not the correct interpretation of the nest species. Users may choose to look at these data more closely if investigating brood parasitism.

b: Flag/remove attempts with Outcome of "no breeding behavior observed" (u3), "inactive" (i), and "not monitored" (n). These data likely represent either nests from a previous season or nests that were previously active this season, but no active dates can be attributed to. Users may choose to include these data if looking at habitat or location data without measures of phenology or success.

c: Flag/remove attempts with Outcome of "invasive spp management" (f5). In these attempts, the participant chose to remove/alter eggs or nests of invasive species. If the user is interested in analyzing participants' habits in invasive spp management or rates of brood parasitism, they should consider not using this method. Note: "Invasive spp management", may also be incorrectly chosen by participants trying to indicate nest failure by an invasive species where a failure code should have been used.

d: Flag/remove attempts if outcome is a fail code (f, f1, f2, f3, f6, f7) but recorded fledged host young > 0. In this case, a participant may have either characterized the nest's outcome incorrectly, recorded the presence of fledged host young incorrectly, or mischaracterized brood parasite young as host, but discerning the true outcome may not be possible. If an attempt produces any number of fledged host young, the attempt was successful.

e: Flag/remove attempt if Outcome is success (s1), but recorded fledged host young = 0. Inverse of d.

f: Flag/remove attempt if # hatched host young > clutch size. This may indicate incorrectly entered data or long lengths of time between nest checks where summary data was not properly updated. A user may choose to review these attempts by looking with caution at the nest visit data to validate hatched young and clutch sizes.

g: Flag/remove attempt if # fledged host young > # hatched host young. Similar to f.

h: Flag/remove attempts that occur in international waters Subnational.Code == "XX-". Consider not using this procedure if analyzing coastal or water-nesting species (eg. Osprey nesting on channel markers). Subnational code is automatically assigned based on coordinates supplied by the participant. Many attempts identified as XX- likely resulted from incorrect or inaccurate coordinates.

i: Flag attempts with impossible nesting periods. Incorrect years are sometimes entered between date summary fields by participants. This may produce impossibly long nesting periods. To account for nesting phenologies in all hemispheres, this procedure identifies attempts in which (1) Fledge Date - First Lay Date > 365 days (Snowy Albatross nesting period), (2) Hatch Date - First Lay Date > 84 days (Southern Brown Kiwi incubation period), or (3) Fledge Date - Hatch Date > 300 days (Snowy Albatross nestling period). These dates represent the maximum nest phonological period for any bird species and are not realistic for the majority of the NestWatch dataset. We encourage users to determine reasonable phenologies for their species of interest and use nw.filterphenology to run a finer filter on nest phenology dates.

j: Flag/remove attempts where the # days between the first and the last visit are > 365 days. Additional check to identify nest attempts where year portion of dates between nest visits were likely incorrect entered. A user may choose to review these attempts individually to verify if a typo occurred.

Examples

# Load example wren data
wrens <- nestwatchR::wren_quickstart

# Flag data not meeting procedures e, f, g, and h
nw.cleandata(data = wrens, mode = "flag",
             methods = c("a", "b", "c", "d", "e", "f", "g", "h"))
#> ... Beginning to identify nesting attempts that do not meet the criteria. This may take a minute ...
#> ... Identified nesting attempts have been noted with 'FLAGGED' in the new dataset in column 'Flagged.Attempt'.

# Remove data not meeting procedures i or j.
nw.cleandata(data = wrens, mode = "remove",
             methods = c("i", "j"))
#> ... Beginning to identify nesting attempts that do not meet the criteria. This may take a minute ...
#> ... Identified nesting attempts have been removed from the new dataset.